Use a YAML for preferred_entries
Lots of jankiness in this implementation. In particular, the for-loop
in get_preferred_entry_map seems like it could be replced with
something better. Also, I had to change the signature on best_entry,
even though all of its arguments are borrowed immutably, which made me
sad. Maybe ask the Rust devs about this.