Kill "{,old }variant of" in defs list
If a definition starts with "variant of" or "old variant of", it means
"people sometimes write the entry word when they intend to write this
other, more obscure word". I don't think the student really needs to
know these, and having them in the defs list often makes the word
easier to guess.