--traditional flag to put trad characters first
The --traditional flag makes traditional characters appear first on
"English" and "Pinyin" cards. The "Traditional" card will also appear
before the "Simplified" card when studying, and the "Taiwan Pinyin", in
the few cases where it's present, will appear before the (mainland)
"Pinyin" on all cards. Cards will be counted as pinyin duplicates if
either of one card's pronunciations matches either of the other's.